requirements.txt
asgiref==3.5.2
astroid==2.12.10
black @ git+https://github.com/psf/black@b60b85b234d6a575f636d0a125478115f993c90c
click==8.1.3
dill==0.3.5.1
Django==4.1.1
django-phonenumber-field==7.0.0
fontawesomefree==6.2.0
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
phonenumbers==8.12.56
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.4
pylint==2.15.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.4
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
whitenoise==6.2.0
wrapt==1.14.1
